issue705: make MaximumCommonType into a property of NodeWithVariables
