+++ javaparser-core/src/main/java/com/github/javaparser/printer/PrettyPrintVisitor.java
@@ -395,7 +395,7 @@ public class PrettyPrintVisitor implements VoidVisitor<Void> {
         printMemberAnnotations(n.getAnnotations(), arg);
         printModifiers(n.getModifiers());
         if (!n.getVariables().isEmpty()) {
-            getMaximumCommonType(n).accept(this, arg);
+            n.getMaximumCommonType().accept(this, arg);
         }
 
         printer.print(" ");
@@ -415,7 +415,7 @@ public class PrettyPrintVisitor implements VoidVisitor<Void> {
         printJavaComment(n.getComment(), arg);
         n.getName().accept(this, arg);
 
-        Type commonType = getMaximumCommonType(n.getAncestorOfType(NodeWithVariables.class).get());
+        Type commonType = n.getAncestorOfType(NodeWithVariables.class).get().getMaximumCommonType();
 
         Type type = n.getType();
 
@@ -805,7 +805,7 @@ public class PrettyPrintVisitor implements VoidVisitor<Void> {
         printModifiers(n.getModifiers());
 
         if (!n.getVariables().isEmpty()) {
-            getMaximumCommonType(n).accept(this, arg);
+            n.getMaximumCommonType().accept(this, arg);
         }
         printer.print(" ");
 
@@ -1403,58 +1403,6 @@ public class PrettyPrintVisitor implements VoidVisitor<Void> {
         for (int i = 0; i < commentsAtEnd; i++) {
             everything.get(everything.size() - commentsAtEnd + i).accept(this, null);
         }
-    }
-
-    /**
-     * Returns the type that maximum shared type between all variables.
-     * The minimum common type does never include annotations on the array level.
-     * <p>
-     * <br/>For <code>int a;</code> this is int.
-     * <br/>For <code>int a,b,c,d;</code> this is also int.
-     * <br/>For <code>int a,b[],c;</code> this is also int.
-     * * <br/>For <code>int[] a[][],b[],c[][];</code> this is int[][].
-     * <p>
-     * Visible for testing.
-     */
-    static Type getMaximumCommonType(NodeWithVariables<?> nodeWithVariables) {
-        // we use a local class because we cannot use an helper static method in an interface
-        class Helper {
-            // Conceptually: given a type we start from the Element Type and get as many array levels as indicated
-            // From the implementation point of view we start from the actual type and we remove how many array
-            // levels as needed to get the target level of arrays
-            // It returns null if the type has less array levels then the desired target
-            private Type toArrayLevel(Type type, int level) {
-                if (level > type.getArrayLevel()) {
-                    return null;
-                }
-                for (int i = type.getArrayLevel(); i > level; i--) {
-                    type = ((ArrayType) type).getComponentType();
-                }
-                return type;
-            }
-        }
-
-        Helper helper = new Helper();
-        int level = 0;
-        boolean keepGoing = true;
-        // In practice we want to check for how many levels of arrays all the variables have the same type,
-        // including also the annotations
-        while (keepGoing) {
-            final int currentLevel = level;
-            // Now, given that equality on nodes consider the position the simplest way is to compare
-            // the pretty-printed string got for a node. We just check all them are the same and if they
-            // are we just just is not null
-            Object[] values = nodeWithVariables.getVariables().stream().map(v -> {
-                Type t = helper.toArrayLevel(v.getType(), currentLevel);
-                return t == null ? null : t.toString();
-            }).distinct().toArray();
-            if (values.length == 1 && values[0] != null) {
-                level++;
-            } else {
-                keepGoing = false;
-            }
-        }
-        return helper.toArrayLevel(nodeWithVariables.getVariables().get(0).getType(), --level);
     }
 
 }
